Update for the latest dependencies, fix issues and tests, and add github actions support (#1)
* Add go module files

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>

* Add reaper option for a PreDeleteFn

Add reaper option for a PreDeleteFn. This function runs before an
expired session is deleted. This provide the library user an extra hook
to do additional work before deleting a session (e.g., revoke tokens to
a remote endpoint).

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>

* Fix formatting errors

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>

* Fix unit tests and load deadlock

The load unit-test hanged indefinitely. Presumably, this is because of a
deadlock caused by starting a write transaction inside a read
transaction. From the boltdb docs:

> Read-only transactions and read-write transactions should not depend on
> one another and generally shouldn't be opened simultaneously in the
> same goroutine. This can cause a deadlock as the read-write transaction
> needs to periodically re-map the data file but it cannot do so while a
> read-only transaction is open.

In addition, the protobuf error message returned for a case of the load
unit test seems to have changed, so amended to the latest text.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>

* fixup! Add go module files

* fixup! Add go module files

* change module name, change bolt to bbolt

* Add support for github actions

* Update the docs

Co-authored-by: Yannis Zarkadas <yanniszark@arrikto.com>